# backend/ms-geolocalisation/src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: PGVPT - MS Géolocalisation API
  description: Propriétaire des coordonnées, du référentiel administratif, des zones géographiques et touristiques, des points
    d'accès et des recherches spatiales. Une géolocalisation cible exactement un patrimoine ou une entreprise; les références
    sont des UUID.
  version: 14.0.0
  contact:
    name: PGVPT
    email: contact@pgvpt.sn
servers:
  - url: http://localhost:3002/api/v1/geolocalisations
    description: Environnement local
  - url: https://api.pgvpt.sn/geolocalisation/api/v1
    description: Environnement de production
...
paths:
...
  /pays:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les pays
      operationId: getAllPays
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des pays
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePays'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer un pays
      operationId: createPays
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaysCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pays'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /pays/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir un pays
      operationId: getPays
      responses:
        '200':
          description: Pays trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pays'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier un pays
      operationId: updatePays
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaysUpdate'
      responses:
        '200':
          description: Pays modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pays'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer un pays
      operationId: deletePays
      responses:
        '204':
          description: Pays supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/PaysId'
  /regions:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les régions
      operationId: getRegions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des régions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageRegion'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer une region
      operationId: createRegion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /regions/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir une region
      operationId: getRegion
      responses:
        '200':
          description: Region trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier une région
      operationId: updateRegion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionUpdate'
      responses:
        '200':
          description: Region modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer une région
      operationId: deleteRegion
      responses:
        '204':
          description: Region supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/RegionId'
  /departements:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les départements
      operationId: getDepartements
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des départements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDepartement'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer un(e) departement
      operationId: createDepartement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartementCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Departement'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /departements/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir un departement
      operationId: getDepartement
      responses:
        '200':
          description: Departement trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Departement'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier un département
      operationId: updateDepartement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartementUpdate'
      responses:
        '200':
          description: Département modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Departement'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer un département
      operationId: deleteDepartement
      responses:
        '204':
          description: Département supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/DepartementId'
  /communes:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les communes
      operationId: getCommunes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des communes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCommune'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer une commune
      operationId: createCommune
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommuneCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commune'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /communes/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir une commune
      operationId: getCommune
      responses:
        '200':
          description: Commune trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commune'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier une commune
      operationId: updateCommune
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommuneUpdate'
      responses:
        '200':
          description: Commune modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commune'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer une commune
      operationId: deleteCommune
      responses:
        '204':
          description: Commune supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/CommuneId'
  /villages:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les villages
      operationId: getVillages
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des villages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVillage'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer un village
      operationId: createVillage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillageCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Village'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /villages/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir un village
      operationId: getVillage
      responses:
        '200':
          description: Village trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Village'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier un village
      operationId: updateVillage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillageUpdate'
      responses:
        '200':
          description: Village modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Village'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer un village
      operationId: deleteVillage
      responses:
        '204':
          description: Village supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/VillageId'
  /quartiers:
    get:
      tags:
        - Referentiel administratif
      summary: Lister les quartiers
      operationId: getQuartiers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste des quartiers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageQuartier'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Referentiel administratif
      summary: Créer un quartier
      operationId: createQuartier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuartierCreate'
      responses:
        '201':
          description: Ressource créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quartier'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      parameters: []
      security:
        - bearerAuth: []
  /quartiers/{id}:
    get:
      tags:
        - Referentiel administratif
      summary: Obtenir un quartie
      operationId: getQuartier
      responses:
        '200':
          description: Village trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quartier'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Referentiel administratif
      summary: Modifier un quartier
      operationId: updateQuartier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuartierUpdate'
      responses:
        '200':
          description: Quartier modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quartier'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Referentiel administratif
      summary: Supprimer un quartier
      operationId: deleteQuartier
      responses:
        '204':
          description: Quartier supprimée
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []
    parameters:
      - $ref: '#/components/parameters/QuartierId'
...
components:
  parameters:
    PaysId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RegionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DepartementId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CommuneId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VillageId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: Numéro de page, indexé à partir de 0.
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Nombre maximal d’éléments par page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Unauthorized:
      description: Authentification requise
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    Pays:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - nom
        - code
        - devise
      properties:
        nom:
          type: string
          minLength: 1
          maxLength: 150
          example: Sénégal
        code:
          type: string
          minLength: 2
          maxLength: 3
          example: SEN
        codeIso2:
          type: string
          minLength: 2
          maxLength: 2
          example: SN
        codeIso3:
          type: string
          minLength: 3
          maxLength: 3
          example: SEN
        devise:
          type: string
          example: Franc CFA BCEAO
        codeDevise:
          type: string
          minLength: 3
          maxLength: 3
          example: XOF
        indicatifTelephonique:
          type: string
          example: '+221'
        langueOfficielle:
          type: string
          example: Français
        continent:
          type: string
          example: Afrique
        actif:
          type: boolean
          default: true
    PaysCreate:
      type: object
      required:
        - nom
        - code
        - devise
      properties:
        nom:
          type: string
          minLength: 1
          maxLength: 150
          example: Sénégal
        code:
          type: string
          minLength: 2
          maxLength: 3
          example: SEN
        codeIso2:
          type: string
          minLength: 2
          maxLength: 2
          example: SN
        codeIso3:
          type: string
          minLength: 3
          maxLength: 3
          example: SEN
        devise:
          type: string
          example: Franc CFA BCEAO
        codeDevise:
          type: string
          minLength: 3
          maxLength: 3
          example: XOF
        indicatifTelephonique:
          type: string
          example: '+221'
        langueOfficielle:
          type: string
          example: Français
        continent:
          type: string
          example: Afrique
        actif:
          type: boolean
          default: true
    PaysUpdate:
      type: object
      required:
        - nom
        - code
        - devise
      properties:
        nom:
          type: string
          minLength: 1
          maxLength: 150
          example: Sénégal
        code:
          type: string
          minLength: 2
          maxLength: 3
          example: SEN
        codeIso2:
          type: string
          minLength: 2
          maxLength: 2
          example: SN
        codeIso3:
          type: string
          minLength: 3
          maxLength: 3
          example: SEN
        devise:
          type: string
          example: Franc CFA BCEAO
        codeDevise:
          type: string
          minLength: 3
          maxLength: 3
          example: XOF
        indicatifTelephonique:
          type: string
          example: '+221'
        langueOfficielle:
          type: string
          example: Français
        continent:
          type: string
          example: Afrique
        actif:
          type: boolean
          default: true
    Region:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - code
        - nom
      properties:
        pays:
          $ref: '#/components/schemas/PaysReference'
        code:
          type: string
          example: DK
        nom:
          type: string
          example: Dakar
        codePostal:
          type: string
          example: '10000'
        description:
          type: string
    RegionCreate:
      type: object
      required:
        - paysId
        - code
        - nom
      properties:
        paysId:
          type: string
          format: uuid
        code:
          type: string
          example: DK
        nom:
          type: string
          example: Dakar
        codePostal:
          type: string
          example: '10000'
        description:
          type: string
    RegionUpdate:
      type: object
      required:
        - paysId
        - code
        - nom
      properties:
        paysId:
          type: string
          format: uuid
        code:
          type: string
          example: DK
        nom:
          type: string
          example: Dakar
        codePostal:
          type: string
          example: '10000'
        description:
          type: string
    Departement:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - code
        - nom
      properties:
        region:
          $ref: '#/components/schemas/RegionReference'
        code:
          type: string
          example: DKR
        nom:
          type: string
          example: Dakar
        description:
          type: string
    DepartementCreate:
      type: object
      required:
        - regionId
        - code
        - nom
      properties:
        regionId:
          type: string
          format: uuid
        code:
          type: string
          example: DKR
        nom:
          type: string
          example: Dakar
        description:
          type: string
    DepartementUpdate:
      type: object
      required:
        - regionId
        - code
        - nom
      properties:
        regionId:
          type: string
          format: uuid
        code:
          type: string
          example: DKR
        nom:
          type: string
          example: Dakar
        description:
          type: string
    Commune:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
      required:
        - id
        - code
        - nom
      properties:
        departement:
          $ref: '#/components/schemas/DepartementReference'
        code:
          type: string
          example: GOR
        nom:
          type: string
          example: Gorée
        description:
          type: string
    CommuneCreate:
      type: object
      required:
        - departementId
        - code
        - nom
      properties:
        departementId:
          type: string
          format: uuid
        code:
          type: string
          example: GOR
        nom:
          type: string
          example: Gorée
        description:
          type: string
    CommuneUpdate:
      type: object
      required:
        - departementId
        - code
        - nom
      properties:
        departementId:
          type: string
          format: uuid
        code:
          type: string
          example: GOR
        nom:
          type: string
          example: Gorée
        description:
          type: string
...
    PagePays:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Pays'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    PageRegion:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Region'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    PageDepartement:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Departement'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
    PageCommune:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Commune'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []